adding a class that resembles rails ActiveRecord
This class stores a json file. The object is an array of
dictionaries, same principle as with MongoDB.

With

class A < StatusPage::ActiveRecord
  self.reload_cache()
end

a new storage class is created.

To get all entries:

  A.all

To find entries based on a query

  A.find(query)

where query can be

- dictionary: { key1: val1, key2: val2 }
  the keys must exist and the values must match

- StatusPage::Query

StatusPage::Query is a simple query system. It contains of this classes:

  StatusPage::AndQuery
  StatusPage::OrQuery
  StatusPage::NotQuery

For every class there are shortcuts

  _QAND    (equivalent to StatusPage::AndQuery.new)
  _QOR
  _QNOT

Example:

get all entries where "x" is 3, "y" is 4 and NOT(OR("a" is 1 or "b" is  2))

  A.find(x: 3, y:4, "_not" => _QNOT(_QOR(a:1, b:2)))

A new record can be created like this:

 a = A.new

or

 a = A.new val1: 1, val2: 3, val3: 5

setter & getters are automatically added to the key/pair values.

 a = A.new val1: 1, val2: 3, val3: 5
 a.val1 = a.val2 + a.val3

 # adding a new value
 a.val4 = a.val1 * 10

 # removing a value
 a.val3 = nil

 # saving object
 a.save!